Implement measurements of the master boot record
This patch adds support for measuring the boot block of the MBR and logging the
measurement. It also puts an 'event' separator into the log that can then be
seen in Linux's /sys/kernel/security/tpm0/ascii_bios_measurements. More
low-level C functions are added for measuring and logging of disk related data,
along with their FORTH-level counterparts.

Logging follows the specifications found on the following page:

http://www.trustedcomputinggroup.org/resources/pc_client_work_group_specific_implementation_specification_for_conventional_bios

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>