slof: Make linker script variables accessible
Make linker script variables related to 'text' addresses available
to the code so we can measure the static core root of trust contents.

When hashing the 'data' part of SLOF we do not end up with the same
measurements for the same firmware when booting with different
configurations, so we don't make those available.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>