tcgbios: Implement tpm_hash_log_extend_event_file
Implement tpm_hash_log_extend_event_file() that allows to measure
the contents of a file into a given PCR and log it with the
given event type and description. The caller may choose to have
the size of the original ELF image detected so that only data
from the ELF image are hashed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>